quanticaiengineering/requirements.txt
# Production dependencies
# Python: >=3.10

# Web Framework
Flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.3
gunicorn==21.2.0

# Groq — LLM generation
groq>=0.13.0

# Embeddings (local, lightweight)
sentence-transformers>=2.2.0

# Vector store
faiss-cpu==1.13.0

# Numerics
numpy>=1.26.0

# Document parsing
beautifulsoup4>=4.12.0
markdown>=3.5.0
pypdf>=4.0.0

# Environment & configuration
python-dotenv>=1.0.0
pydantic>=2.6.0

# Utilities
tenacity>=8.2.0
tqdm>=4.66.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
regex>=2023.12.25
click>=8.1.0
rich>=13.7.0
